continued implementing new memory system
