+++ libs/componentModels/src/common/CMakeLists.txt
@@ -2,4 +2,5 @@ TARGET_SOURCES(SWEVAL_BACKENDS_LIB PRIVATE
     StaticBranchPredictModel.cpp
     DynamicBranchPredictModel.cpp
     ConfigurableMemoryModel.cpp
+    ConfigurableMemoryModel/MemoryInstanceManager.cpp
 )
